server/requirements.txt
boto3
glob2
prisma
fastapi[standard]
genai
uvicorn
google-genai
importlib_metadata
passlib[bcrypt]
python-jose[cryptography]
python-dotenv